+++ smart-recipe-book/src/pages/Logged-Out/TheFeed.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import NavLoggedOut from '../../components/Logged-Out/NavLoggedOut';
-import Filter from '../../components/Filter';
+import Filter from '../../components/Logged-Out/Filter';
 
 function TheFeed() {
